+++ LookupAnything/Framework/Subjects/FarmerSubject.cs
@@ -7,6 +7,7 @@ using Pathoschild.Stardew.LookupAnything.Framework.Fields;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Objects;
+using SFarmer = StardewValley.Farmer;
 
 namespace Pathoschild.Stardew.LookupAnything.Framework.Subjects
 {
@@ -17,7 +18,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Subjects
         ** Properties
         *********/
         /// <summary>The lookup target.</summary>
-        private readonly Farmer Target;
+        private readonly SFarmer Target;
 
         /// <summary>Simplifies access to private game code.</summary>
         private readonly IReflectionHelper Reflection;
@@ -29,7 +30,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Subjects
         /// <summary>Construct an instance.</summary>
         /// <param name="farmer">The lookup target.</param>
         /// <param name="reflectionHelper">Simplifies access to private game code.</param>
-        public FarmerSubject(Farmer farmer, IReflectionHelper reflectionHelper)
+        public FarmerSubject(SFarmer farmer, IReflectionHelper reflectionHelper)
             : base(farmer.Name, null, "Player")
         {
             this.Target = farmer;
@@ -40,7 +41,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Subjects
         /// <param name="metadata">Provides metadata that's not available from the game data directly.</param>
         public override IEnumerable<ICustomField> GetData(Metadata metadata)
         {
-            Farmer farmer = this.Target;
+            SFarmer farmer = this.Target;
 
             int maxSkillPoints = metadata.Constants.PlayerMaxSkillPoints;
             int[] skillPointsPerLevel = metadata.Constants.PlayerSkillPointsPerLevel;
@@ -48,11 +49,11 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Subjects
             yield return new GenericField("Farm name", farmer.farmName);
             yield return new GenericField("Favourite thing", farmer.favoriteThing);
             yield return new GenericField("Spouse", farmer.spouse, hasValue: farmer.isMarried());
-            yield return new SkillBarField("Farming skill", farmer.experiencePoints[Farmer.farmingSkill], maxSkillPoints, skillPointsPerLevel);
-            yield return new SkillBarField("Mining skill", farmer.experiencePoints[Farmer.miningSkill], maxSkillPoints, skillPointsPerLevel);
-            yield return new SkillBarField("Foraging skill", farmer.experiencePoints[Farmer.foragingSkill], maxSkillPoints, skillPointsPerLevel);
-            yield return new SkillBarField("Fishing skill", farmer.experiencePoints[Farmer.fishingSkill], maxSkillPoints, skillPointsPerLevel);
-            yield return new SkillBarField("Combat skill", farmer.experiencePoints[Farmer.combatSkill], maxSkillPoints, skillPointsPerLevel);
+            yield return new SkillBarField("Farming skill", farmer.experiencePoints[SFarmer.farmingSkill], maxSkillPoints, skillPointsPerLevel);
+            yield return new SkillBarField("Mining skill", farmer.experiencePoints[SFarmer.miningSkill], maxSkillPoints, skillPointsPerLevel);
+            yield return new SkillBarField("Foraging skill", farmer.experiencePoints[SFarmer.foragingSkill], maxSkillPoints, skillPointsPerLevel);
+            yield return new SkillBarField("Fishing skill", farmer.experiencePoints[SFarmer.fishingSkill], maxSkillPoints, skillPointsPerLevel);
+            yield return new SkillBarField("Combat skill", farmer.experiencePoints[SFarmer.combatSkill], maxSkillPoints, skillPointsPerLevel);
             yield return new GenericField("Luck", $"{this.GetSpiritLuckMessage()}{Environment.NewLine}({(Game1.dailyLuck >= 0 ? "+" : "")}{Game1.dailyLuck * 100}% to many random checks)");
         }
 
@@ -60,7 +61,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Subjects
         /// <param name="metadata">Provides metadata that's not available from the game data directly.</param>
         public override IEnumerable<IDebugField> GetDebugFields(Metadata metadata)
         {
-            Farmer target = this.Target;
+            SFarmer target = this.Target;
 
             // pinned fields
             yield return new GenericDebugField("immunity", target.immunity, pinned: true);
@@ -79,7 +80,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Subjects
         /// <returns>Returns <c>true</c> if a portrait was drawn, else <c>false</c>.</returns>
         public override bool DrawPortrait(SpriteBatch spriteBatch, Vector2 position, Vector2 size)
         {
-            Farmer farmer = this.Target;
+            SFarmer farmer = this.Target;
             FarmerSprite sprite = farmer.FarmerSprite;
 
             farmer.FarmerRenderer.draw(spriteBatch, sprite.CurrentAnimationFrame, sprite.CurrentFrame, sprite.SourceRect, position, Vector2.Zero, 0.8f, Color.White, 0, 1f, farmer);